+++ CHANGELOG.md
@@ -17,6 +17,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added Features and Improvements 🙌:
 - Allows selecting 0.05kg/st/lb steps for more precision
 - Major performance improvements by making app async
+- Allow using imperial units for entering height
 
 ### Bugfix 🐛:
 - Fix broken calendar when choosing custom first Day of week option, #417 #418